Fix bug and add security

+++ app/controllers/AdminPage.java
@@ -10,74 +10,74 @@ import java.util.*;
 
 public class AdminPage extends Controller {
 
-    private static User _user = User.findByUserId(Long.parseLong(session("userId")));
-
     @Security.Authenticated(Secured.class)
     public static Result index() {
+        User _user = User.findByUserId(Long.parseLong(session("userId")));
+        if(checkPermission(_user)){
+            flash("voting_result_close","access denied");
+            return redirect(routes.ProjectList.index());
+        }
         Logger.info("["+_user.username+"] user admin page.");
         List<Rate> rates = Rate.findAll();
         List<User> users = User.findAll();
         List<Settings> webconfig = Settings.findAll();
-
-        if(_user.idtype == User.ADMINISTRATOR) {
-            return ok(adminpage.render(users, Project.findAll(), rates, webconfig,RateCriterion.findAll(),VoteCriterion.findAll()));
-        } else {
-            return redirect(routes.ProjectList.index());
-        }
+        return ok(adminpage.render(users, Project.findAll(), rates, webconfig,RateCriterion.findAll(),VoteCriterion.findAll()));
     }
 
     @Security.Authenticated(Secured.class)
     public static Result addUser(){
-        if(_user.idtype == User.ADMINISTRATOR) {
-            User user = Form.form(User.class).bindFromRequest().get();
-            user.save();
-            Logger.info("["+_user.username+"] add new user.("+user.id+")");
-            return redirect(routes.AdminPage.index()+"#users");
-        } else {
+        User _user = User.findByUserId(Long.parseLong(session("userId")));
+        if(checkPermission(_user)){
             flash("voting_result_close","access denied");
             return redirect(routes.ProjectList.index());
         }
+        User user = Form.form(User.class).bindFromRequest().get();
+        user.save();
+        Logger.info("["+_user.username+"] add new user.("+user.id+")");
+        return redirect(routes.AdminPage.index()+"#users");
     }
 
     @Security.Authenticated(Secured.class)
     public static Result addProject(){
-        if(_user.idtype == User.ADMINISTRATOR) {
-            Project project = Form.form(Project.class).bindFromRequest().get();
-            project.save();
-            Logger.info("["+_user.username+"] add new project.("+project.id+")");
-            return redirect(routes.AdminPage.index()+"#projects");
-        } else {
+        User _user = User.findByUserId(Long.parseLong(session("userId")));
+        if(checkPermission(_user)){
             flash("voting_result_close","access denied");
             return redirect(routes.ProjectList.index());
         }
+        Project project = Form.form(Project.class).bindFromRequest().get();
+        project.save();
+        Logger.info("["+_user.username+"] add new project.("+project.id+")");
+        return redirect(routes.AdminPage.index()+"#projects");
     }
     @Security.Authenticated(Secured.class)
     public static Result addRateCriterion(){
-        if(_user.idtype == User.ADMINISTRATOR) {
-            RateCriterion rateC = Form.form(RateCriterion.class).bindFromRequest().get();
-            rateC.save();
-            Logger.info("[" + _user.username + "] add new Rate Criterion.(" + rateC.id + ")");
-            return redirect(routes.AdminPage.index()+"#criterions");
-        } else {
+        User _user = User.findByUserId(Long.parseLong(session("userId")));
+        if(checkPermission(_user)){
             flash("voting_result_close","access denied");
             return redirect(routes.ProjectList.index());
         }
+        RateCriterion rateC = Form.form(RateCriterion.class).bindFromRequest().get();
+        rateC.save();
+        Logger.info("[" + _user.username + "] add new Rate Criterion.(" + rateC.id + ")");
+        return redirect(routes.AdminPage.index()+"#criterions");
     }
     @Security.Authenticated(Secured.class)
     public static Result addVoteCriterion(){
-        if(_user.idtype == User.ADMINISTRATOR) {
-            VoteCriterion voteC = Form.form(VoteCriterion.class).bindFromRequest().get();
-            voteC.save();
-            Logger.info("[" + _user.username + "] add new Vote Criterion.(" + voteC.id + ")");
-            return redirect(routes.AdminPage.index()+"#criterions");
-        } else {
+        User _user = User.findByUserId(Long.parseLong(session("userId")));
+        if(checkPermission(_user)){
             flash("voting_result_close","access denied");
             return redirect(routes.ProjectList.index());
         }
+        VoteCriterion voteC = Form.form(VoteCriterion.class).bindFromRequest().get();
+        voteC.save();
+        Logger.info("[" + _user.username + "] add new Vote Criterion.(" + voteC.id + ")");
+        return redirect(routes.AdminPage.index()+"#criterions");
     }
     @Security.Authenticated(Secured.class)
     public static Result deleteUsers(){
-        if(_user.idtype != User.ADMINISTRATOR) {
+        User _user = User.findByUserId(Long.parseLong(session("userId")));
+        if(checkPermission(_user)){
+            flash("voting_result_close","access denied");
             return redirect(routes.ProjectList.index());
         }
         Map<String, String[]> map = request().body().asFormUrlEncoded();
@@ -107,6 +107,11 @@ public class AdminPage extends Controller {
     }
     @Security.Authenticated(Secured.class)
     public static Result deleteRate(Long id){
+        User _user = User.findByUserId(Long.parseLong(session("userId")));
+        if(checkPermission(_user)){
+            flash("voting_result_close","access denied");
+            return redirect(routes.ProjectList.index());
+        }
         Rate rate = Rate.findById(id);
         Logger.info("[" + _user.username + "] delete rate.(" + rate.id + ")("+rate.projectId+")");
         rate.delete();
@@ -114,7 +119,9 @@ public class AdminPage extends Controller {
     }
     @Security.Authenticated(Secured.class)
     public static Result deleteRateCriterion(Long id){
-        if(_user.idtype != User.ADMINISTRATOR) {
+        User _user = User.findByUserId(Long.parseLong(session("userId")));
+        if(checkPermission(_user)){
+            flash("voting_result_close","access denied");
             return redirect(routes.ProjectList.index());
         }
         RateCriterion rateCs = RateCriterion.findById(id);
@@ -131,7 +138,9 @@ public class AdminPage extends Controller {
     }
     @Security.Authenticated(Secured.class)
     public static Result deleteVoteCriterion(Long id){
-        if(_user.idtype != User.ADMINISTRATOR) {
+        User _user = User.findByUserId(Long.parseLong(session("userId")));
+        if(checkPermission(_user)){
+            flash("voting_result_close","access denied");
             return redirect(routes.ProjectList.index());
         }
         VoteCriterion voteCs = VoteCriterion.findById(id);
@@ -149,6 +158,11 @@ public class AdminPage extends Controller {
 
     @Security.Authenticated(Secured.class)
     public static Result editUser(){
+        User _user = User.findByUserId(Long.parseLong(session("userId")));
+        if(checkPermission(_user)){
+            flash("voting_result_close","access denied");
+            return redirect(routes.ProjectList.index());
+        }
         User newuser = Form.form(User.class).bindFromRequest().get();
         User olduser = User.findByUserId(newuser.id);
         olduser.firstname = newuser.firstname;
@@ -163,11 +177,19 @@ public class AdminPage extends Controller {
     }
     @Security.Authenticated(Secured.class)
     public static Result saveSetting(){
+        User _user = User.findByUserId(Long.parseLong(session("userId")));
+        if(checkPermission(_user)){
+            flash("voting_result_close","access denied");
+            return redirect(routes.ProjectList.index());
+        }
         DynamicForm dynamicForm = new DynamicForm().bindFromRequest();
         List<Settings> settings = Settings.findAll();
         for(Settings item : settings){
             Settings.update(item.keyName, dynamicForm.get(item.keyName));
         }
         return redirect(routes.AdminPage.index()+"#configs");
+    }
+    public static boolean checkPermission(User _user){
+        return _user.idtype != User.ADMINISTRATOR ;
     }
 }
